refactor(sync): promote LinkDeviceThisDeviceCard to standalone widget (#388)
Promotes the second of the three private cards in
link_device_screen.dart to a standalone widget under
presentation/widgets/. Follows #447 (HowItWorksCard) and continues
the cleanup started by #417.

The card displays the device's anonymous user id with a copy-to-
clipboard button. Pulling it out:
- Drops 78 lines of inline private widget from the screen
- Eliminates 2 imports the analyzer flagged as unused once the
  widget left the file (services.dart, snackbar_helper.dart)
- Lets the copy interaction be exercised by widget tests

link_device_screen.dart: 232 → 152 lines (–80).

Regression tests in `link_device_this_device_card_test.dart`
(5 tests):
1. renders the device id when one is provided
2. renders "Not connected" when myId is null
3. shows the copy button when myId is non-null
4. hides the copy button when myId is null
5. tapping the copy button writes the id to the clipboard
   (uses a mocked SystemChannels.platform handler)

All 110 existing sync tests still pass.

Refs #388
