+++ lib/features/sync/presentation/screens/link_device_screen.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../../core/sync/supabase_client.dart';
-import '../../../../core/widgets/snackbar_helper.dart';
 import '../../../../l10n/app_localizations.dart';
 import '../../providers/link_device_provider.dart';
 import '../widgets/link_device_how_it_works_card.dart';
+import '../widgets/link_device_this_device_card.dart';
 
 class LinkDeviceScreen extends ConsumerStatefulWidget {
   const LinkDeviceScreen({super.key});
@@ -39,7 +38,7 @@ class _LinkDeviceScreenState extends ConsumerState<LinkDeviceScreen> {
       body: ListView(
         padding: const EdgeInsets.all(16),
         children: [
-          _ThisDeviceCard(myId: _myId),
+          LinkDeviceThisDeviceCard(myId: _myId),
           const SizedBox(height: 16),
           _ImportFromDeviceCard(codeController: _codeController),
           const SizedBox(height: 16),
@@ -50,85 +49,6 @@ class _LinkDeviceScreenState extends ConsumerState<LinkDeviceScreen> {
   }
 }
 
-/// Card showing the current device's anonymous user id with a copy button.
-class _ThisDeviceCard extends StatelessWidget {
-  final String? myId;
-
-  const _ThisDeviceCard({required this.myId});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final l10n = AppLocalizations.of(context);
-
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                const Icon(Icons.smartphone, size: 20),
-                const SizedBox(width: 8),
-                Text(
-                  l10n?.linkDeviceThisDeviceLabel ?? 'This device',
-                  style: theme.textTheme.titleMedium
-                      ?.copyWith(fontWeight: FontWeight.bold),
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            Text(
-              l10n?.linkDeviceShareCodeHint ??
-                  'Share this code with your other device:',
-              style: theme.textTheme.bodySmall
-                  ?.copyWith(color: theme.colorScheme.onSurfaceVariant),
-            ),
-            const SizedBox(height: 8),
-            Container(
-              padding:
-                  const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-              decoration: BoxDecoration(
-                color: theme.colorScheme.surfaceContainerHighest,
-                borderRadius: BorderRadius.circular(8),
-              ),
-              child: Row(
-                children: [
-                  Expanded(
-                    child: SelectableText(
-                      myId ??
-                          (l10n?.linkDeviceNotConnected ?? 'Not connected'),
-                      style: theme.textTheme.bodySmall
-                          ?.copyWith(fontFamily: 'monospace'),
-                    ),
-                  ),
-                  if (myId != null)
-                    IconButton(
-                      icon: const Icon(Icons.copy, size: 18),
-                      tooltip:
-                          l10n?.linkDeviceCopyCodeTooltip ?? 'Copy code',
-                      onPressed: () {
-                        Clipboard.setData(ClipboardData(text: myId!));
-                        SnackBarHelper.show(
-                          context,
-                          AppLocalizations.of(context)?.deviceCodeCopied ??
-                              'Device code copied',
-                        );
-                      },
-                      padding: EdgeInsets.zero,
-                      constraints: const BoxConstraints(
-                          minWidth: 32, minHeight: 32),
-                    ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
 
 /// Card with the code input field and the "Import data" button. Needs
 /// access to the parent's TextEditingController (kept local to the widget
